+++ packages/aptos-sdk/src/nft/mint.test.ts
@@ -1,14 +1,15 @@
-import { createTestAptosState } from "../common/test"
+import { createTestAptosState, createV1Token } from "../common/test"
 import { AptosNft } from "./nft"
 
 describe("mint nft", () => {
-  const { aptos, config, wallet } = createTestAptosState()
+  const state = createTestAptosState()
+  const { aptos, config, wallet } = state
   const mintClass = new AptosNft(aptos, wallet, config)
 
-  test.skip("mint by mintDigitalAssetTransaction function", async () => {
+  test("mint by mintDigitalAssetTransaction function", async () => {
     const uri = "ipfs://QmWYpMyoaUGNRSQbwhw97xM8tcRWm4Et598qtzmzsau7ch/"
     const { tx, tokenAddress } = await mintClass.mintWithCollectionName({
-      collectionName: "Test collection #42601579",
+      collectionName: "Test collection #3072715",
       name: "Mytoken #4",
       description: "Description of Mytoken #4",
       uri,
@@ -26,6 +27,11 @@ describe("mint nft", () => {
       description: "Description of Mytoken #4",
       uri,
     })
+    console.log("tx", JSON.stringify(tx, null, "  "))
+  })
+
+  test("mint v1", async () => {
+    const tx = await createV1Token(state)
     console.log("tx", JSON.stringify(tx, null, "  "))
   })
 })