+++ packages/aptos-sdk/src/nft/transfer.test.ts
@@ -1,5 +1,5 @@
 import { Account } from "@aptos-labs/ts-sdk"
-import { createTestAptosState, mintTestToken } from "../common/test"
+import { createTestAptosState, createV1Token, mintTestToken } from "../common/test"
 import { AptosNft } from "./nft"
 
 describe("transfer nft", () => {
@@ -10,12 +10,26 @@ describe("transfer nft", () => {
   test("transfer", async () => {
     const recepientAccount = Account.generate()
     const receipentAddress = recepientAccount.accountAddress.toStringLong()
-    const testTokenAddress = await mintTestToken(state)
+    const { tokenAddress } = await mintTestToken(state)
 
-    await transferClass.transfer(testTokenAddress, receipentAddress)
+    await transferClass.transfer(tokenAddress, receipentAddress)
 
     const assets = await aptos.getOwnedDigitalAssets({ ownerAddress: receipentAddress })
-    const tokenOfNewOwner = assets.find(asset => asset.token_data_id === testTokenAddress)
+    const tokenOfNewOwner = assets.find(asset => asset.token_data_id === tokenAddress)
     expect(tokenOfNewOwner).toBeTruthy()
+  })
+
+  test("transfer v1 token", async () => {
+    const recepientAccount = Account.generate()
+    const receipentAddress = recepientAccount.accountAddress.toStringLong()
+    const { propertyVersion, collectionName, tokenName, creator } = await createV1Token(state)
+    await transferClass.transferV1Token(
+      receipentAddress,
+      creator,
+      collectionName,
+      tokenName,
+      propertyVersion.toString(),
+      "1",
+    )
   })
 })