+++ packages/sdk/src/sdk-blockchains/aptos/listing.ts
@@ -1,6 +1,7 @@
 import type { AptosSdk, SupportedNetwork as SupportedAptosNetwork } from "@rarible/aptos-sdk"
 import type { IBlockchainTransaction } from "@rarible/sdk-transaction"
 import type { OrderId } from "@rarible/api-client"
+import { Blockchain } from "@rarible/api-client"
 import { Action } from "@rarible/action"
 import { extractId } from "@rarible/sdk-common"
 import { BlockchainAptosTransaction } from "@rarible/sdk-transaction"
@@ -15,7 +16,7 @@ import type * as OrderCommon from "../../types/order/common"
 import { MaxFeesBasePointSupport, OriginFeeSupport, PayoutsSupport } from "../../types/order/fill/domain"
 import { getNftContractAddress, getOrderId } from "../../common/utils"
 import { getCurrencyAssetType } from "../../common/get-currency-asset-type"
-import { convertAptosToUnionOrderId, getFeeObject, getSupportedCurrencies } from "./common"
+import { convertAptosToUnionOrderId, getSupportedCurrencies } from "./common"
 
 export class AptosListing {
   constructor(
@@ -34,24 +35,51 @@ export class AptosListing {
       id: "send-tx" as const,
       run: async (request: OrderCommon.OrderInternalRequest) => {
         const aptosItemId = extractId(request.itemId)
+        const aptosItem = await this.apis.item.getItemById({ itemId: `${Blockchain.APTOS}:${aptosItemId}` })
+        const aptosCollection = await this.apis.collection.getCollectionById({
+          collection: aptosItem.collection!,
+        })
+
         if (request.originFees && request.originFees.length > 1) {
           throw new Error("Origin fees should consist only 1 item")
         }
+        const feeObject = await this.sdk.order.getFeeObject(
+          request.originFees?.length
+            ? {
+                address: extractId(request.originFees[0].account),
+                value: request.originFees[0].value,
+              }
+            : undefined,
+        )
         const assetType = getCurrencyAssetType(request.currency)
         if (assetType["@type"] !== "CURRENCY_NATIVE") {
           throw new Error("Only native token currency is available for sell operation")
         }
         const startTime = Math.floor(Date.now() / 1000)
-        const objectAddress = await this.sdk.order.sell(
-          aptosItemId,
-          await getFeeObject({
-            originFees: request.originFees || [],
-            defaultFeeAddress: this.sdk.order.getFeeScheduleAddress(),
-            createFeeSchedule: this.sdk.order.createFeeSchedule,
-          }),
-          startTime,
-          toBn(request.price.toString()).multipliedBy(APT_DIVIDER).toFixed(),
-        )
+
+        let objectAddress
+        if (aptosCollection.extra?.standard === "v1") {
+          if (!aptosItem.extra) {
+            throw new Error("No extra field in API item")
+          }
+          const firstCreator = extractId(aptosItem.creators[0].account)
+          objectAddress = await this.sdk.order.sellV1(
+            feeObject,
+            firstCreator,
+            aptosItem.extra.onChainCollectionName,
+            aptosItem.extra.onChainTokenName,
+            aptosItem.extra.propertyVersionV1,
+            startTime,
+            toBn(request.price.toString()).multipliedBy(APT_DIVIDER).toFixed(),
+          )
+        } else {
+          objectAddress = await this.sdk.order.sell(
+            aptosItemId,
+            feeObject,
+            startTime,
+            toBn(request.price.toString()).multipliedBy(APT_DIVIDER).toFixed(),
+          )
+        }
         return convertAptosToUnionOrderId(objectAddress)
       },
     })