+++ packages/sdk/src/types/ethereum/domain.ts
@@ -1,5 +1,7 @@
 import type { IBlockchainTransaction } from "@rarible/sdk-transaction"
 import type { Action } from "@rarible/action"
+import type { GetOrderBuyTxRequest } from "@rarible/protocol-ethereum-sdk/build/order/fill-order/types"
+import type { TransactionData } from "@rarible/protocol-ethereum-sdk/build/order/fill-order/types"
 
 export type CryptopunkWrapRequest = {
   punkId: number
@@ -11,3 +13,5 @@ export type CryptopunkUnwrapRequest = {
 
 export type ICryptopunkWrap = Action<"approve-tx" | "wrap-tx", CryptopunkWrapRequest, IBlockchainTransaction>
 export type ICryptopunkUnwrap = Action<"unwrap-tx", CryptopunkUnwrapRequest, IBlockchainTransaction>
+
+export type IGetBuyTxData = (request: GetOrderBuyTxRequest) => Promise<TransactionData>