+++ CHANGELOG.md
@@ -3,6 +3,51 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.13.68-fix.49](https://github.com/rarible/sdk/compare/v0.13.68-fix.48...v0.13.68-fix.49) (2024-09-04)
+
+### Bug Fixes
+
+- barbie collection cadence 1.0 update ([3a6c1d7](https://github.com/rarible/sdk/commit/3a6c1d73c3d64283a1dcad3ce116e00675cbbc68))
+
+## [0.13.68-fix.48](https://github.com/rarible/sdk/compare/v0.13.68-fix.47...v0.13.68-fix.48) (2024-09-03)
+
+### Bug Fixes
+
+- flow cadence update ([ea6b614](https://github.com/rarible/sdk/commit/ea6b614f25c76576dcb718dfe5000865abdbdb6b))
+
+## [0.13.68-fix.47](https://github.com/rarible/sdk/compare/v0.13.68-fix.46...v0.13.68-fix.47) (2024-09-02)
+
+### Bug Fixes
+
+- **PLTF-552:** opensea order fix ([48c4aeb](https://github.com/rarible/sdk/commit/48c4aeb1d63ac56616185ed425284d5751437818))
+
+## [0.13.68-fix.46](https://github.com/rarible/sdk/compare/v0.13.68-fix.45...v0.13.68-fix.46) (2024-08-26)
+
+### Bug Fixes
+
+- **PLTF-573:** logging return data ([6876737](https://github.com/rarible/sdk/commit/6876737bdf9122f1bdbedd67be2bd3227d5f58b6))
+
+## [0.13.68-fix.45](https://github.com/rarible/sdk/compare/v0.13.68-fix.44...v0.13.68-fix.45) (2024-08-20)
+
+### Bug Fixes
+
+- aptos fee address improvement ([7fb0d64](https://github.com/rarible/sdk/commit/7fb0d64cc2d891608c26370838b3686a4bea98a2))
+
+## [0.13.68-fix.44](https://github.com/rarible/sdk/compare/v0.13.68-fix.43...v0.13.68-fix.44) (2024-08-13)
+
+### Features
+
+- aptos zero fee addr ([23c18f7](https://github.com/rarible/sdk/commit/23c18f7a6741ca44ee246bf502bad77273a9f355))
+
+## [0.13.68-fix.43](https://github.com/rarible/sdk/compare/v0.13.68-fix.42...v0.13.68-fix.43) (2024-08-12)
+
+### Features
+
+- add todo ([1727c2d](https://github.com/rarible/sdk/commit/1727c2d7c183fe49eb71ccbfc9916db690b41db3))
+- aptos v1 orders ([08c33f1](https://github.com/rarible/sdk/commit/08c33f1fee0f0d7a1c2d6ff36dbc9691193326e8))
+- introduce sellV1 method ([7c84a7f](https://github.com/rarible/sdk/commit/7c84a7fd883415236860b0092d87efc1b7c730e4))
+- v1 aptos token support ([8cdb9c7](https://github.com/rarible/sdk/commit/8cdb9c7b2eff2f8a792d5c13e63c85adc56d2970))
+
 ## [0.13.68-fix.42](https://github.com/rarible/sdk/compare/v0.13.68-fix.41...v0.13.68-fix.42) (2024-07-29)
 
 ### Features